Bug 1988032. Don't async scroll with anchors if there is an active view transition. r=layout-reviewers,emilio
Being inside an active view transition forces a scroll frame to be inactive.

https://searchfox.org/firefox-main/rev/70425199e9a5fa80aa7419fa51763013a67226e1/layout/generic/ScrollContainerFrame.cpp#4349

IsInViewTransitionCapture is set on a stacking context basis. This means that when painting walks from a placeholder frame to its out of flow then IsInViewTransitionCapture follows. But the ASR tree (which we must construct for async scrolling the anchor pos) strictly follows the parent (not the placeholder). So we'd have to do a completely separate walk, and I'm not sure it's worth it. If it turns out to be important we can implement something, or even better just remove that code in DecideScrollableLayer which I'm pretty sure we don't need but I don't want to open that can of worms right now.

Differential Revision: https://phabricator.services.mozilla.com/D273593

UltraBlame original commit: aa8819fe8168fdd75031530a028244abd75130af