Bug 1988030. Allow passing a limit ancestor to DisplayPortUtils::OneStepInASRChain that we don't walk past. r=hiro,layout-reviewers
With CSS anchor pos we know that the anchor is in the containing block of the anchored content. We use that to limit a walk up the frame tree in subsequent patches and make the computation simpler. So we need OneStepInASRChain to obey that limit too.

Differential Revision: https://phabricator.services.mozilla.com/D271804

UltraBlame original commit: 6d1306b4671cb2e093fe209044494273b6ec83c0