Bug 1988030. Detect and reject an anchor that is outside of the positioned frame's containing block. r=layout-reviewers,emilio
This should be the case but due to spec issues that we are still getting sorted out in bug 1997026 it does not currently hold with our existing code that accurately implemented an older version of the spec. The anchor pos async scroll code in this bug depends on this condition even if other anchor pos code seems to be okay without it as of now.

Differential Revision: https://phabricator.services.mozilla.com/D273580

UltraBlame original commit: fa7ec7d8ecb523428094a0437658ed229001b32b