Bug 1989686 - Don't look for minidumps in tests that don't care about them r=afranchuk,extension-reviewers,robwu,mconley
This is a workaround for the intermittent failures when generating
minidumps on macOS 10.15 in two browser mochitests. The tests crash
several frames but don't care about the crash reports, however they
currently fail because sometimes a minidump is not generated. This patch
instructs the harness to not try and clean up the minidumps, this
prevents the spurious failures and can be removed once we make macOS
minidump generation more robust (bug 1614931).

Differential Revision: https://phabricator.services.mozilla.com/D274061

UltraBlame original commit: f724d041273237cf2c7347051d7b30c955c2229f