Bug 2002715 - Upgrade rustc-dev to 1.82.0. r=firefox-build-system-reviewers,sergesanspaille
In bug bug 1945020, we upgraded rustc-dev to a version before 1.82.0
because our patches for -Z build-std didn't apply or work with 1.82.
Subsequently, we applied patches to enable features that were stabilized
in final 1.82 but weren't yet in the nightly we had picked.

After multiple dead-ends trying to keep our patches working similarly to
how they used to, it just dawned on me that we didn't need to try to
handle things on the cargo end anymore:

Up to rustc 1.81, cargo would use a virtual workspace for -Z build-std.
As of 1.82, it uses a real workspace with its own Cargo.toml. So instead
of having cargo programmatically handle the situation, we can... just
edit that Cargo.toml instead.

For the record, a .cargo/config.toml with the vendoring configuration
that `cargo vendor` outputs does *not* work.

Differential Revision: https://phabricator.services.mozilla.com/D274363

UltraBlame original commit: 84a5f17e080745223c35dba123414e4075699571