Bug 2001825 - Make `EditorBase::GetPreferredIMEState()` return `Result<IMEState, nsresult>` r=m_kato
For making the implementation is for which editor, the method should be
a pure virtual function and `TextEditor` and `HTMLEditor` implement it
as they needed.

Additionally, it should return `Result` to avoid the caller to use
fallback IME state even temporarily.

Differential Revision: https://phabricator.services.mozilla.com/D274386

UltraBlame original commit: 051325f945177a6a2ded39ed02874d83993619c3