Bug 2001073 - Fix another scrolling jitter issue with fractional scrolling enabled. r=gfx-reviewers,lsalzman
This partially undoes the fix in #1999575, but retains an adjusted
reftest from that patch. The separate fix in #1971253 is actually
the correct fix for this problem, and fixes both of those bugs,
without needing the change in #1999575. Add an additional reftest for
the specific case in this bug.

Differential Revision: https://phabricator.services.mozilla.com/D274881

UltraBlame original commit: b4d167623b6273ac7ff1181826da22833525630c