Bug 1999440 - Rename existing layout.css.scroll-behavior.* prefs to layout.css.scroll-snap.*. r=hiro,layout-reviewers,perftest-reviewers,aglavic
These prefs are no longer used for scroll-behavior if
layout.css.scroll-behavior.same-physics-as-user-input=true,
which will be the default in the future.

They are always used for scroll-snap, so it's more accurate to
name the prefs after scroll-snap.

The patch also updates tests which deliberately tweak these prefs to
control the speed of a scroll animation governed by scroll-behavior
(e.g. to slow it down), to also set
layout.css.scroll-behavior.same-physics-as-user-input=false,
so that the prefs continue to have their intended effect on the test.

Differential Revision: https://phabricator.services.mozilla.com/D272350

UltraBlame original commit: ed2caba32fe2169d2e32e3dd56bf1673f8597095