Bug 2000077 - sqlite persistence for tab notes r=jswinarton,tabbrowser-reviewers
Basic persistence of tab notes in a new tabnotes.sqlite database in the profile root directory.

Currently has a single table `tabnotes` with a canonical URL, note text, and creation timestamp. The TabNotesController manages the startup/shutdown cycle for the TabNotes singleton, while TabNotes talks to the database directly.

We only have an async database connection available on the main thread, so this forced the TabNotes API to become async. This is a little awkward in some cases like the tab context menu where things normally run synchronously, but hopefully there aren't too many race conditions.

This patch doesn't include any optimizations for URL matching. I was planning to store a hashed value of the canonical URL in each row in order to use the same optimizations available in places.sqlite, but I found that the places code registered a custom C++ hashing function with sqlite for use in the places database. As a result, I think optimizations should be tackled in a separate bug(s).

Differential Revision: https://phabricator.services.mozilla.com/D275025

UltraBlame original commit: 1307a66912694bb503796d1ecb3d41e92212ff5d