Bug 2004166 - Ensure a continuation is an abspos containing block in PushAbsoluteContainingBlock(). r=dholbert,layout-reviewers
Assume an ib-split `-moz-block-inside-inline-wrapper` is split by a column-span.
The wrapper **before** column-span might not be an absolute containing block if
it does not have any abspos children. In that case, `nsIFrame::Init()` won't
create an absolute containing block for the wrapper continuation **after** the
column-span.

This patch removes the assertions and properly sets up an absolute containing
block for the continuation when it needs one.

Differential Revision: https://phabricator.services.mozilla.com/D275275

UltraBlame original commit: cf4ea2cf0c475ac4a7d199f0918babac20d5f674