Bug 2000988 - Get rid of `handled` in `nsRange::CutContents()` r=dom-core,smaug
It's initialized at start of the loop without referring it. Therefore,
it can be a local variable in the loop.  However, it's set after the
last referrer.  So, we can delete the last set.

On the other hand, the name is really general one.  Therefore,
developers may want to refer it after the last referrer even though
nobody won't set it properly after the last referrer.

The reason why the flag is required is, the flow needs to go to the
fallback path if it's not in the specific cases.  However, the most
cases are illegal so that we can assert or continue in such illegal
cases.

Differential Revision: https://phabricator.services.mozilla.com/D273120

UltraBlame original commit: 4819c326fdf2c10b00d4d13d25654e9dd57546db