Bug 1873716: Simplify the fingerprinting content blocking log logic r=timhuang
Presently we use CheckSuspiciousFingerprintingActivity to see
if the number of fingerprinting activities exceed a theshold
and if so, we indicate that an entry has suspicious fingerprinting.

That threshold is 1. And we don't have any plans to adjust it.

So we can just remove the check altogether: if that function is
called, it is called immediately after creating or finding a log
of the two relevant types: Canvas or Font fingerprinting.  So it
will _always_ return true.

SKIP_BMO_CHECK

Differential Revision: https://phabricator.services.mozilla.com/D273659

UltraBlame original commit: b80938513b4669274e570a52c14f3bdb1555a57a