Bug 1873716: Handle about:blank iframes that do fingerprinting r=timhuang
We encountered a fingerprinter that created an about:blank iframe
and did fingerprinting there. This iframe had no channel, so it
did not get a ContentBlockingEvent on it.  Traverse the document
tree upwards to find a document we can log it on.

SKIP_BMO_CHECK

Differential Revision: https://phabricator.services.mozilla.com/D274380

UltraBlame original commit: 5e55a1f34d89f89be0d6323036db95c28dae990c