Bug 2004912 - Forward declare nsHttpResponseHead in SharedSubResourceCache.h r=emilio,firefox-style-system-reviewers
Making changes to nsHttpResponseHead.h or nsHttpHeaderArray.h needs almost a
full recompilation because of the inclusion of the header in
SharedSubResourceCache.h which is then included in many other places.

This patch moves the destructor impl to the cpp file so the
definition of nsHttpResponseHead is accessible to the
destructor via UniquePtr.

Differential Revision: https://phabricator.services.mozilla.com/D275602

UltraBlame original commit: 00ea22fb70da26981e34507cd2f6bebcdcbb6555