Bug 2005092 - Wait for RemoteMediaDataEncoderChild::Construct to resolve before handling Init. r=media-playback-reviewers,karlt
Init cannot be handled before Construct has completed. For synchronously
created encoders, Init may get called first, but in that case, we just
need to cache the request until we resolve a promise from Construct.
This race is possible because we defer Construct until the encoding
process promise resolves (it may still be launching). Simply checking
the promise like we did before was insufficient because the Init call
might win the race against the process launch as well.

This patch makes it so that we never send the Init message until the
response has been received for the Construct message.

Differential Revision: https://phabricator.services.mozilla.com/D275700

UltraBlame original commit: 3bc86c8a35a25c089360470f7633ba245dbe196c